TestLogin

+++ DungeonMaker/Assets/Scripts/Network/Packet.cs
@@ -69,7 +69,7 @@ public enum ClientPackets
         /// <summary>Inserts the length of the packet's content at the start of the buffer.</summary>
         public void WriteLength()
         {
-            buffer.InsertRange(0, BitConverter.GetBytes(buffer.Count)); // Insert the byte length of the packet at the very beginning
+            buffer.InsertRange(0, BitConverter.GetBytes((short) buffer.Count)); // Insert the byte length of the packet at the very beginning
         }
 
         /// <summary>Inserts the given int at the start of the buffer.</summary>
